# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing volleyball games played with friends.",
        "title": "Volleyball Tracker API",
        "contact": {},
        "version": "2.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/players": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "List Players",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cscercel_volleyball-tracker_internal_db.Player"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Create Player",
                "parameters": [
                    {
                        "description": "Player Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cscercel_volleyball-tracker_internal_db.Player"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/players/leaderboard": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Get Leaderboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match Type",
                        "name": "match_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Season",
                        "name": "season",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/github_com_cscercel_volleyball-tracker_internal_db.GetLeaderboardRow"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/players/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Get Player By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Match Type",
                        "name": "match_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Season",
                        "name": "season",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cscercel_volleyball-tracker_internal_db.GetPlayerStatsByIDRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the name of an existing player by their UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Update Player Name",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Player UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Player Name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cscercel_volleyball-tracker_internal_db.Player"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Delete Player",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Player UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/players/{id}/history": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Get Player Match History",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Match Type",
                        "name": "match_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Season",
                        "name": "season",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/github_com_cscercel_volleyball-tracker_internal_db.GetPlayerSeasonalMatchesRow"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_cscercel_volleyball-tracker_internal_db.GetLeaderboardRow": {
            "type": "object",
            "properties": {
                "conceded": {
                    "type": "integer"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "efficiency_rate": {},
                "id": {
                    "type": "string"
                },
                "longest_streak": {
                    "type": "integer"
                },
                "losses": {
                    "type": "integer"
                },
                "match_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "otl": {
                    "type": "integer"
                },
                "played": {
                    "type": "integer"
                },
                "player_id": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "scored": {
                    "type": "integer"
                },
                "season": {
                    "type": "integer"
                },
                "streak": {
                    "type": "integer"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "win_rate": {},
                "wins": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_cscercel_volleyball-tracker_internal_db.GetPlayerSeasonalMatchesRow": {
            "type": "object",
            "properties": {
                "blue_score": {
                    "type": "integer"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "id": {
                    "type": "string"
                },
                "match_type": {
                    "type": "string"
                },
                "red_score": {
                    "type": "integer"
                },
                "season": {
                    "type": "integer"
                }
            }
        },
        "github_com_cscercel_volleyball-tracker_internal_db.GetPlayerStatsByIDRow": {
            "type": "object",
            "properties": {
                "conceded": {
                    "type": "integer"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "efficiency_rate": {},
                "id": {
                    "type": "string"
                },
                "longest_streak": {
                    "type": "integer"
                },
                "losses": {
                    "type": "integer"
                },
                "match_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "otl": {
                    "type": "integer"
                },
                "played": {
                    "type": "integer"
                },
                "player_id": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "scored": {
                    "type": "integer"
                },
                "season": {
                    "type": "integer"
                },
                "streak": {
                    "type": "integer"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "win_rate": {},
                "wins": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_cscercel_volleyball-tracker_internal_db.Player": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
        "pgtype.Timestamptz": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}